fix: Add existing instance check

+++ models/classes/resources/Service/ClassDeleter.php
@@ -167,6 +167,10 @@ class ClassDeleter implements ClassDeleterInterface
                 continue;
             }
 
+            if (!$instance->exists()) {
+                continue;
+            }
+
             try {
                 $this->resourceRepository->delete(
                     new ResourceRepositoryContext(
